fix todolist
