+++ src/api/config.ts
@@ -1,0 +1,6 @@
+const configs = {
+    TODO_URL: process.env.REACT_APP_TODO_URL,
+    AUTH_URL: process.env.REACT_APP_AUTH_URL
+}
+
+export default configs;
