+++ src/features/auth/components/InputTag/InputTag.css
@@ -1,27 +1,3 @@
-html {
-    width: 100%;
-    height: 100%;
-}
-
-.login-form, .register-form {
-    margin: auto;
-    margin-top: 160px;
-    width: 400px;
-    border: 1px solid #e0e0e0;
-    border-radius: 10px;
-    padding:  40px 40px 40px 40px;
-    box-sizing: border-box;
-}
-
-.login-form > p, .register-form > p {
-    font-size: 24px;
-    font-weight: 600px;
-}
-
-img[alt='logo'] {
-    width: 80px;
-}
-
 input {
     box-sizing: border-box;
     width: 100%;
@@ -65,10 +41,4 @@ input[name="submit"] {
     border: none;
     color: white;
     margin: 40px 0 20px 0;
-}
-
-p.to-resgister, p.to-login {
-    font-size: 14px;
-    margin-block-start: 0;
-    margin-block-end: 0;
 }
