+++ src/features/auth/constants/auth.endpoints.ts
@@ -1,0 +1,4 @@
+export enum AuthEndPointsEnum {
+    REGISTER = '/register',
+    LOGIN = '/login'
+}
