--- src/features/todolist/api/config.ts
@@ -1,6 +1,0 @@
-const configs = {
-    TODO_URL: 'https://627a92f04a5ef80e2c1d395f.mockapi.io/',
-    AUTH_URL: 'https://reqres.in/api/'
-}
-
-export default configs;
