+++ src/features/todolist/api/todolist.api.ts
@@ -1,106 +1,31 @@
-import { createAsyncThunk} from '@reduxjs/toolkit';
-import { TodoStatus, TodoType } from '../types/types';
-import configs from './config';
-import { ResponseTodoType } from './../types/types';
+import { TodolistEndPointsEnum } from '../constants/todolist.endpoints';
+import { todoAPI } from './../../../api/api';
+import { TodoType, createNewTodoType, ResponseTodoType } from './../types/todolist.types';
 
-const axios = require('axios');
+const getTodosApi = (): Promise<ResponseTodoType> => {
+    return todoAPI.get(TodolistEndPointsEnum.TODOS)
+}
 
-const todoAPI = axios.create({
-    baseURL: configs.TODO_URL,
-});
+const createNewTodoApi = (link: string, data: createNewTodoType): Promise<ResponseTodoType> => {
+    return todoAPI.post(link, data);
+}
 
-export const getTodoList = createAsyncThunk(
-    'todos/getTodoList',
-    async () => {
-        try {
-            const response = await todoAPI
-                .get('todos')
-                .then((res: ResponseTodoType) => res)
-            if (response.status === 200 || response.status === 201){
-                return response.data;
-            }
-        }
-        catch(error) {
-            alert('Có lỗi xảy ra!');
-        }
-    }
-)
+const editTodoApi = (apiUrl: string, data: TodoType): Promise<ResponseTodoType> => {
+    return todoAPI.put(apiUrl, data);
+}
 
-export const creatNewTodo = createAsyncThunk(
-    'todos/creatNewTodo',
-    async (todoTitle: string, { dispatch }) => {
-        try {
-            const response = await todoAPI
-                .post('todos', {
-                    title: todoTitle,
-                    status: TodoStatus.inProgress
-                })
-                .then((res: ResponseTodoType) => res)
-            if (response.status === 200 || response.status === 201) {
-                dispatch(getTodoList());
-            }
-        }
-        catch(error) {
-            alert('Thêm mới không thành công!');
-        }
-    }
-)
+const updateTodoApi = (apiUrl: string, data: TodoType): Promise<ResponseTodoType> => {
+    return todoAPI.put(apiUrl, data);
+}
 
-export const editTodo = createAsyncThunk(
-    'todos/editTodo',
-    async (todo: TodoType, { dispatch }) => {
-        try {
-            const response = await todoAPI
-                .put(`todos/${todo.id}`,
-                    {
-                        title: todo.title
-                    }
-                )
-                .then((res: ResponseTodoType) => res)
-            if (response.status === 200 || response.status === 201) {
-                dispatch(getTodoList());
-            }
-        }
-        catch(error) {
-            alert('Chỉnh sửa không thành công!');
-        }
-    }
-)
+const deleteTodoApi = (apiUrl: string): Promise<ResponseTodoType> => {
+    return todoAPI.delete(apiUrl);
+}
 
-export const updateTodo = createAsyncThunk(
-    'todos/updateTodo',
-    async (todo: TodoType, { dispatch }) => {
-        try {
-            const response = await todoAPI
-                .put(`todos/${todo.id}`,
-                    {
-                        status: (todo.status === TodoStatus.inProgress) ? TodoStatus.done : TodoStatus.inProgress
-                    }
-                )
-                .then((res: ResponseTodoType) => res)
-            if (response.status === 200 || response.status === 201) {
-                dispatch(getTodoList());
-            }
-        }
-        catch(error) {
-            alert('Update không thành công!');
-        }
-    }
-)
-
-export const deleteTodo = createAsyncThunk(
-    'todos/deleteTodo',
-    async (todoId: number, { dispatch }) => {
-        try {
-            const response = await todoAPI
-                .delete(`todos/${todoId}`)
-                .then((res: ResponseTodoType) => res)
-            if (response.status === 200 || response.status === 201) {
-                dispatch(getTodoList());
-            }
-        }
-        catch(error) {
-            alert('Xóa không thành công!');
-        }
-    }
-)
+export const todoListApi = {
+    getTodosApi,
+    createNewTodoApi,
+    editTodoApi,
+    updateTodoApi,
+    deleteTodoApi
+}
