+++ src/features/todolist/components/AddTodo/AddTodo.css
@@ -19,7 +19,10 @@
 .add-todo button {
     width: 100px;
     height: 40px;
+    color: white;
     background-color: #53b953;
+    border-radius: 3px;
+    border: none;
 }
 
 .add-todo button:hover {
