+++ src/features/todolist/constants/todolist.endpoints.ts
@@ -1,0 +1,3 @@
+export enum TodolistEndPointsEnum {
+    TODOS = 'todos'
+}
