+++ src/features/todolist/constants/todolist.enums.ts
@@ -1,0 +1,4 @@
+export enum TodoStatus {
+    inProgress = 'In Progress',
+    done = 'Done'
+}
